memoクラスのファイル切り分け
